set default route, Shop, ShopItem components, create cart with useState
This commit sets the default route to /home. This is triggered during
a useEffect when the page is initially loaded. The user will
automatically be redirected to /home. A Shop page has been added and
a Shop item component has been added which will represent each of
the items available in the shop. useState has been used in
RouteSwitch.js to create an inital empty cart. This has additionally
been passed to the navbar so the current number of items in the cart can
be displayed.