add a cart system using state, add 'add to cart' button functionality
This commit adds a cart system using state which keeps track of items
users add to their cart. The quantity of an item added to cart can be
adjusted by the user by editing the input field next to the item.
I also have noticed that there is a problem with the way I am
route switching. It causes the cart state to be lost when the page
redirects the user with href within the header button link tags.