Fix cart display on Shop.js
Fixed an issue where the cart container div wasn't correctly
updating state. It turns out that I was incorrectly updating state.
The concat method had to be used instead of directly updating the state
with (a copy of the array but updated). The array was still
a reference to the intial array causing react not to update the
component.