add item imgs to assets, add quantity input validation
This commit adds the shop item images to assets however does not
actually add them to the shop page yet. The quantity input for shop
items has been updated to require a number between 1 and 9. The user
and press the arrows on the input form to increase or decrease the
quantity added to cart. Alternatively the user can directly type in
the desired quantity as long as it is within the acceptable range.