rename asset imgs, add images to shop items, add checkout button/form
This commit changes the names of the notebooks.jpg to notebook.jpg and
changes sharpener.jpg to pencilSharpener.jpg to be more accurate to
what is shown in the image. This commit also adds images to both
the checkout page and to the shop page. On the checkout page there is
also now a form containing a checkout button. When pressed the user will
be redirected to the homepage simulating a sucessful page checkout.