Add checkout success page, add checkout items
This commit adds a checkout success page which is shown after
the user submits their cart to checkout. There is also now a
checkout item container which holds all the items in the users cart.
This container is shown at /checkout. The checkout item container
does not allow changing of item quantity through an input element.